fix: standardized the offer page with the new activity page
