+++ .env
@@ -10,7 +10,7 @@ DATABASE_URL=postgres://postgres:password@localhost:5432/holaplex
 SOLANA_ENDPOINT=https://holaplex-main-9e4a.mainnet.rpcpool.com
 NEXT_PUBLIC_SOLANA_ENDPOINT=https://holaplex-main-9e4a.mainnet.rpcpool.com
 NEXT_PUBLIC_INDEXER_RPC_URL=https://metaplex-indexer-staging.herokuapp.com
-NEXT_PUBLIC_INDEXER_GRAPHQL_URL="https://graph-test.holaplex.com/v1"
+NEXT_PUBLIC_INDEXER_GRAPHQL_URL="https://graph.holaplex.com/v1"
 NEXT_PUBLIC_ENVIRONMENT=production
 NEXT_PUBLIC_HASURA_GRAPHQL_URL=https://dev-holaplex.hasura.app/v1/graphql
 HASURA_GRAPHQL_ADMIN_SECRET=
