+++ src/pages/profiles/[publicKey]/offers.tsx
@@ -4,9 +4,16 @@ import {
   getProfileServerSideProps,
   WalletDependantPageProps,
 } from '@/views/profiles/getProfileServerSideProps';
-import { ProfileDataProvider } from 'src/views/profiles/ProfileDataProvider';
+import { ProfileDataProvider, useProfileData } from 'src/views/profiles/ProfileDataProvider';
 import { useMemo, useState } from 'react';
-import { useOffersPageQuery } from '@/graphql/indexerTypes';
+import {
+  OffersPageQuery,
+  useOffersPageQuery,
+  // Offer,
+  // Nft as INft,
+  // AhListing,
+  // Marketplace,
+} from '@/graphql/indexerTypes';
 import {
   HOLAPLEX_MARKETPLACE_ADDRESS,
   HOLAPLEX_MARKETPLACE_SUBDOMAIN,
@@ -14,6 +21,7 @@ import {
 import { imgOpt } from '@/lib/utils';
 import Link from 'next/link';
 import { AhListing, Marketplace, Nft, Offer } from '@holaplex/marketplace-js-sdk';
+
 import { DisplaySOL } from '@/components/CurrencyHelpers';
 import { format as formatTime } from 'timeago.js';
 import Button from '@/components/Button';
@@ -22,6 +30,7 @@ import UpdateOfferForm from '@/components/UpdateOfferForm';
 import { useWallet } from '@solana/wallet-adapter-react';
 import Modal from '@/components/Modal';
 import ProfileLayout from '@/views/profiles/ProfileLayout';
+import { ActivityCard } from '@/components/ActivityCard';
 
 enum OfferFilters {
   ALL,
@@ -29,12 +38,15 @@ enum OfferFilters {
   RECEIVED,
 }
 
+// type Offer = OffersPageQuery['ownedNFTs'][0]['offers'][0];
+
 export const getServerSideProps: GetServerSideProps<WalletDependantPageProps> = async (context) =>
   getProfileServerSideProps(context);
 
 const OfferPage = ({ publicKey, ...props }: WalletDependantPageProps) => {
   const [query, setQuery] = useState('');
   const { publicKey: userPK } = useWallet();
+  const { publicKey: profilePK, twitterHandle } = useProfileData();
 
   const [showUpdateOfferModal, setShowUpdateOfferModal] = useState(false);
   const [filter, setFilter] = useState(OfferFilters.ALL);
@@ -48,18 +60,20 @@ const OfferPage = ({ publicKey, ...props }: WalletDependantPageProps) => {
       address: publicKey,
     },
   });
+
   const marketplace = data?.marketplace;
-  const receivedOffers = data?.receivedOffers;
-  const sentOffers = data?.sentOffers;
+  const ownedNfts = data?.ownedNFTs;
+  const nftsWithSentOffers = data?.nftsWithSentOffers;
 
   const receivedCount = useMemo(
-    () => receivedOffers?.reduce((acc, nft) => acc + nft.offers.filter((o) => o).length, 0) || 0,
-    [receivedOffers]
+    () => ownedNfts?.reduce((acc, nft) => acc + nft.offers.filter((o) => o).length, 0) || 0,
+    [ownedNfts]
   );
 
   const sentCount = useMemo(
-    () => sentOffers?.reduce((acc, nft) => acc + nft.offers.filter((o) => o).length, 0) || 0,
-    [sentOffers]
+    () =>
+      nftsWithSentOffers?.reduce((acc, nft) => acc + nft.offers.filter((o) => o).length, 0) || 0,
+    [nftsWithSentOffers]
   );
 
   const offerCount = receivedCount + sentCount;
@@ -95,6 +109,114 @@ const OfferPage = ({ publicKey, ...props }: WalletDependantPageProps) => {
     );
   };
 
+  function getActivityCard(nft: Nft, offer: Offer) {
+    const defaultListing = nft?.listings?.find(
+      (listing) => listing?.auctionHouse?.address.toString() === HOLAPLEX_MARKETPLACE_ADDRESS
+    );
+
+    if (!defaultListing) return null;
+
+    const offerReceived = nft.owner?.address === profilePK;
+
+    return (
+      <ActivityCard
+        key={offer.id}
+        activity={{
+          activityType: 'offer',
+          createdAt: offer.createdAt,
+          id: offer.id,
+          nft: nft as Nft | any,
+          wallets: [
+            {
+              address: offer.buyer as string,
+              twitterHandle: '', // twitterHandle as string,
+            },
+          ],
+          auctionHouse: {
+            address: offer.auctionHouse?.address!,
+            treasuryMint: offer.auctionHouse?.treasuryMint!,
+          },
+          price: offer.price.toNumber(),
+        }}
+        customActionButton={
+          offerReceived ? (
+            !Boolean(offer.buyer === userPK?.toBase58()) ? null : (
+              <div>
+                <Button
+                  onClick={() => {
+                    setCurrNFT(nft as Nft | any);
+                    setShowUpdateOfferModal(true);
+                  }}
+                  secondary
+                  className={`w-full  bg-gray-800 ease-in hover:bg-gray-700`}
+                >
+                  Update offer
+                </Button>
+                <Modal
+                  open={showUpdateOfferModal}
+                  setOpen={setShowUpdateOfferModal}
+                  title={`Update offer`}
+                >
+                  <UpdateOfferForm
+                    listing={defaultListing}
+                    setOpen={setShowUpdateOfferModal}
+                    nft={currNFT as Nft | any}
+                    marketplace={marketplace as Marketplace}
+                    refetch={refetch}
+                    loading={loading}
+                    hasListing={Boolean(defaultListing)}
+                  />
+                </Modal>
+              </div>
+            )
+          ) : (
+            <>
+              {Boolean(nft?.owner?.address === userPK?.toBase58()) && (
+                <AcceptOfferForm
+                  nft={nft as Nft | any}
+                  offer={offer as Offer}
+                  listing={defaultListing as AhListing}
+                  marketplace={marketplace as Marketplace}
+                  refetch={refetch}
+                />
+              )}
+
+              {Boolean(offer.buyer === userPK?.toBase58()) && (
+                <div>
+                  <Button
+                    onClick={() => {
+                      setCurrNFT(nft as Nft | any);
+                      setShowUpdateOfferModal(true);
+                    }}
+                    secondary
+                    className={`w-full bg-gray-800 ease-in hover:bg-gray-700`}
+                  >
+                    Update offer
+                  </Button>
+                  <Modal
+                    open={showUpdateOfferModal}
+                    setOpen={setShowUpdateOfferModal}
+                    title={`Update offer`}
+                  >
+                    <UpdateOfferForm
+                      listing={defaultListing as AhListing}
+                      setOpen={setShowUpdateOfferModal}
+                      nft={currNFT as Nft | any}
+                      marketplace={marketplace as Marketplace}
+                      refetch={refetch}
+                      loading={loading}
+                      hasListing={Boolean(defaultListing)}
+                    />
+                  </Modal>
+                </div>
+              )}
+            </>
+          )
+        }
+      />
+    );
+  }
+
   return (
     <>
       <div className="sticky top-0 z-10 mb-2 flex flex-col items-center gap-6 bg-gray-900 py-4 lg:flex-row lg:justify-between lg:gap-4">
@@ -109,184 +231,22 @@ const OfferPage = ({ publicKey, ...props }: WalletDependantPageProps) => {
         </div>
       </div>
       <div className={`grid grid-cols-1 gap-4`}>
-        {(filter === OfferFilters.ALL || filter === OfferFilters.RECEIVED) &&
-          receivedOffers?.map((receivedOffer) => {
-            const defaultListing = receivedOffer?.listings.find(
-              (listing) =>
-                listing?.auctionHouse?.address.toString() === HOLAPLEX_MARKETPLACE_ADDRESS
-            );
-            return receivedOffer.offers
-              ?.slice()
-              ?.sort(byDate)
-              .map((offer) => (
-                <div
-                  key={offer.id}
-                  className={`flex  flex-row justify-between rounded-lg border border-gray-800 p-4`}
-                >
-                  <div className={`flex items-center justify-start`}>
-                    <Link href={'/nfts/' + receivedOffer.address}>
-                      <a className="flex-shrink-0">
-                        <img
-                          src={imgOpt(receivedOffer?.image, 600)!}
-                          className={`aspect-square h-20 rounded-lg object-cover`}
-                          alt={receivedOffer?.name}
-                        />
-                      </a>
-                    </Link>
-                    <div>
-                      <div className={`ml-4 flex flex-col justify-center`}>
-                        <p className={`mb-0 text-base text-gray-300`}>
-                          <Link href={`/profiles/${offer.buyer}/offers`}>
-                            <a className={`text-white hover:text-gray-300`}>
-                              {offer.buyer === userPK?.toBase58()
-                                ? `You`
-                                : `@${shortenAddress(offer.buyer)}`}
-                            </a>
-                          </Link>{' '}
-                          offered{' '}
-                          <span className={`text-white`}>
-                            <DisplaySOL amount={offer.price} iconVariant={`small`} />
-                          </span>{' '}
-                          for{' '}
-                          <Link href={`/nfts/${receivedOffer.address}`}>
-                            <a className={`text-white hover:text-gray-300`}>{receivedOffer.name}</a>
-                          </Link>
-                        </p>
-                        <p className={`mb-0 mt-2 text-base text-gray-500`}>
-                          {formatTime(offer.createdAt, `en_US`)}
-                        </p>
-                      </div>
-                    </div>
-                  </div>
-                  <div className={`flex items-center`}>
-                    {Boolean(receivedOffer?.owner?.address === userPK?.toBase58()) && (
-                      <AcceptOfferForm
-                        nft={receivedOffer as Nft | any}
-                        offer={offer as Offer}
-                        listing={defaultListing as AhListing}
-                        marketplace={marketplace as Marketplace}
-                        refetch={refetch}
-                      />
-                    )}
-
-                    {Boolean(offer.buyer === userPK?.toBase58()) && (
-                      <div>
-                        <Button
-                          onClick={() => {
-                            setCurrNFT(receivedOffer as Nft | any);
-                            setShowUpdateOfferModal(true);
-                          }}
-                          secondary
-                          className={`bg-gray-800 ease-in hover:bg-gray-700`}
-                        >
-                          Update offer
-                        </Button>
-                        <Modal
-                          open={showUpdateOfferModal}
-                          setOpen={setShowUpdateOfferModal}
-                          title={`Update offer`}
-                        >
-                          <UpdateOfferForm
-                            listing={defaultListing as AhListing}
-                            setOpen={setShowUpdateOfferModal}
-                            nft={currNFT as Nft | any}
-                            marketplace={marketplace as Marketplace}
-                            refetch={refetch}
-                            loading={loading}
-                            hasListing={Boolean(defaultListing)}
-                          />
-                        </Modal>
-                      </div>
-                    )}
-                  </div>
-                </div>
-              ));
-          })}
         {(filter === OfferFilters.ALL || filter === OfferFilters.MADE) &&
-          sentOffers?.map((sentOffer) => {
-            const defaultListing = sentOffer?.listings.find(
-              (listing) =>
-                listing?.auctionHouse?.address.toString() === HOLAPLEX_MARKETPLACE_ADDRESS
-            );
-
-            return sentOffer.offers
+          nftsWithSentOffers?.map((nft) => {
+            return nft.offers
               ?.slice()
               ?.sort(byDate)
-              .map((offer) => (
-                <div
-                  key={offer.id}
-                  className={`flex flex-row flex-wrap justify-between gap-4 rounded-lg border border-gray-800 p-4`}
-                >
-                  <div className={`flex items-center justify-start`}>
-                    <Link href={'/nfts/' + sentOffer.address}>
-                      <a className="flex-shrink-0">
-                        <img
-                          src={imgOpt(sentOffer?.image, 400)!}
-                          className={`aspect-square h-20 rounded-lg`}
-                          alt={sentOffer?.name}
-                        />
-                      </a>
-                    </Link>
-                    <div>
-                      <div className={`ml-4 flex flex-col justify-center`}>
-                        <p className={`mb-0 text-base text-gray-300`}>
-                          <Link href={`/profiles/${offer.buyer}/offers`}>
-                            <a className={`text-white hover:text-gray-300`}>
-                              {offer.buyer === userPK?.toBase58()
-                                ? `You`
-                                : `@${shortenAddress(offer.buyer)}`}{' '}
-                            </a>
-                          </Link>{' '}
-                          offered{' '}
-                          <span className={`text-white`}>
-                            <DisplaySOL amount={offer.price} iconVariant={`small`} />
-                          </span>{' '}
-                          for{' '}
-                          <Link href={`/nfts/${sentOffer.address}`}>
-                            <a className={`text-white hover:text-gray-300`}>{sentOffer.name}</a>
-                          </Link>
-                        </p>
-                        <p className={`mb-0 mt-2 text-base text-gray-500`}>
-                          {formatTime(offer.createdAt, `en_US`)}
-                        </p>
-                      </div>
-                    </div>
-                  </div>
-                  <div className={`flex items-center`}>
-                    {Boolean(offer.buyer === userPK?.toBase58()) && (
-                      <div>
-                        <Button
-                          onClick={() => {
-                            setCurrNFT(sentOffer as Nft | any);
-                            setShowUpdateOfferModal(true);
-                          }}
-                          secondary
-                          className={`w-full  bg-gray-800 ease-in hover:bg-gray-700 md:w-auto `}
-                        >
-                          Update offer
-                        </Button>
-                        <Modal
-                          key={sentOffer.address}
-                          open={showUpdateOfferModal}
-                          setOpen={setShowUpdateOfferModal}
-                          title={`Update offer`}
-                        >
-                          <UpdateOfferForm
-                            listing={defaultListing as AhListing}
-                            setOpen={setShowUpdateOfferModal}
-                            nft={currNFT as Nft | any}
-                            marketplace={marketplace as Marketplace}
-                            refetch={refetch}
-                            loading={loading}
-                            hasListing={Boolean(defaultListing)}
-                          />
-                        </Modal>
-                      </div>
-                    )}
-                  </div>
-                </div>
-              ));
+              .map((offer) => getActivityCard(nft as Nft, offer as Offer));
           })}
+
+        {(filter === OfferFilters.ALL || filter === OfferFilters.RECEIVED) &&
+          ownedNfts?.map((nft) => {
+            return nft.offers
+              ?.slice()
+              ?.sort(byDate)
+              .map((offer) => getActivityCard(nft as Nft, offer as Offer));
+          })}
+
         {offerCount <= 0 && (
           <div>
             <div className={`flex flex-col justify-center rounded-lg border-2 border-gray-800 p-4`}>
